Closes https://github.com/ultralytics/ultralytics/issues/19127

<!--
Thank you 🙏 for your contribution to [Ultralytics](https://www.ultralytics.com/) 🚀! Your effort in enhancing our repositories is greatly appreciated. To streamline the process and assist us in integrating your Pull Request (PR) effectively, please follow these steps:

1. Check for Existing Contributions: Before submitting, kindly explore existing PRs to ensure your contribution is unique and complementary.
2. Link Related Issues: If your PR addresses an open issue, please link it in your submission. This helps us better understand the context and impact of your contribution.
3. Elaborate Your Changes: Clearly articulate the purpose of your PR. Whether it's a bug fix or a new feature, a detailed description aids in a smoother integration process.
4. Ultralytics Contributor License Agreement (CLA): To uphold the quality and integrity of our project, we require all contributors to sign the CLA. Please confirm your agreement by commenting below:

    I have read the CLA Document and I sign the CLA

For more detailed guidance and best practices on contributing, refer to our ✅ [Contributing Guide](https://docs.ultralytics.com/help/contributing/). Your adherence to these guidelines ensures a faster and more effective review process.
--->


## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary  
Improved inference performance and ensured clean environment setup by managing specific environment variables.

### 📊 Key Changes  
- Removed the `CUBLAS_WORKSPACE_CONFIG` environment variable to enhance performance during inference.  
- Cleaned up `PYTHONHASHSEED` and `CUBLAS_WORKSPACE_CONFIG` environment variables in `init_seeds` function for a more deterministic setup.

### 🎯 Purpose & Impact  
- 🏎️ **Performance Boost**: Disabling `CUBLAS_WORKSPACE_CONFIG` avoids unnecessary slowdowns during model inference.  
- 🌱 **Cleaner Environment**: Removing leftover environment variables improves reproducibility and reduces potential issues in multi-session setups.  
- 🤝 **Compatibility**: Ensures better adaptability across different setups by handling environment variables proactively.