@Y-T-G 2025-02-08
Seems like the warnings still appear for tests.

The variable needs to be set before any forward pass. Otherwise, even if it's set, it doesn't get rid of the warning.