@Y-T-G 2025-02-08
Moving it inside model init seems to work