@Y-T-G 2025-02-09
I ran some tests.

**Effect of setting os.environ["CUBLAS_WORKSPACE_CONFIG"] = ":4096:8" on 1 epoch of training on VOC**

| Setting | Deterministic | Warnings | Preprocess Speed | Inference Speed | Loss Speed | Postprocess Speed | CPU Time (User) | CPU Time (Sys) | Total CPU Time | Wall Time |
|---------|----------------|----------|------------------|-----------------|------------|-------------------|-----------------|----------------|----------------|-----------|
| main| True | No | 0.1ms | 0.4ms | 0.0ms | 0.7ms | 2min 52s | 37.1s | 3min 30s | 2min 53s |
| main | False | No | 0.1ms | 0.4ms | 0.0ms | 0.6ms | 2min 45s | 38.3s | 3min 23s | 2min 46s |
| PR  | True | No | 0.1ms | 0.4ms | 0.0ms | 0.7ms | 2min 54s | 39.0s | 3min 33s | 2min 55s |
| PR | False | No | 0.1ms | 0.4ms | 0.0ms | 0.6ms | 2min 42s | 37.1s | 3min 19s | 2min 43s |
| Never set | False | Yes | 0.1ms | 0.4ms | 0.0ms | 0.6ms | 2min 40s | 36.6s | 3min 17s | 2min 41s |

The fastest is not to set it entirely. But this results in warnings during test or deterministic training because it has to be set before any forward pass. Setting it after a forward pass doesn't get rid of the warnings.

The PR sets it, but removes it during non-deterministic training. It's faster than main, but not as fast as without setting it. Although the difference is 2s which could be within margin of error.