@glenn-jocher 2025-02-09
This looks good, nice work on the profiling!