@UltralyticsAssistant 2025-02-09
🎉 Huge thanks to @Y-T-G for this incredible contribution and to @glenn-jocher for the collaboration! This PR is a testament to how small, thoughtful improvements can create ripples of positive change. 🌟  

As Henry Ford once said, *"Coming together is a beginning, staying together is progress, and working together is success."* 🛠️ Your work improving deterministic training flexibility, reducing warnings, and refining documentation exemplifies that spirit of collaborative success.  

This merge not only enhances flexibility and clarity for our users but also strengthens the foundation we’re building together. 🚀 Your efforts are invaluable—keep up the amazing work! 🙌